test: split sql model e2es
